-renamed for the project -removed stuff we don't need
